// frontend/src/types/AdvancedFilter.ts
export type AdvancedFilter = {
  filter:
  |  null
  | 'Maiores preços'
  | 'Menores preços'
  | 'Mais vendidos'
  | 'Mais bem avaliados'
  | 'Maiores gastos'
  | 'Menores gastos'
  | 'Mais comprados'
  | 'Menos comprados'
  | 'Favoritos'
  | 'Menos Favoritos'
  | 'Avaliados'
  | 'Não avaliados'
}